+++ src/equicordplugins/questify/settings/fetching.ts
@@ -4,11 +4,8 @@
  * SPDX-License-Identifier: GPL-3.0-or-later
  */
 
-import type { Quest } from "@vencord/discord-types";
-
 import { fetchAndAlertQuests } from "../utils/fetching";
-import { QL } from "../utils/logging";
-import { getCurrentUserId, getQuestifySettings } from "./access";
+import { getQuestifySettings } from "./access";
 
 let autoFetchInterval: null | ReturnType<typeof setInterval> = null;
 const minimumAutoFetchIntervalValue = 30 * 60; // 30 minutes
@@ -65,31 +62,4 @@ export function stopAutoFetchingQuests(): void {
         clearInterval(autoFetchInterval);
         autoFetchInterval = null;
     }
-}
-
-export function resetQuestsToResume(quest?: Quest, userId?: string): void {
-    const { resumeQuestIDs } = getQuestifySettings();
-    const key = getCurrentUserId(userId);
-
-    if (!key) {
-        QL.warn("RESET_QUESTS_TO_RESUME_NO_USER");
-        return;
-    }
-
-    const resumeState = resumeQuestIDs[key];
-
-    if (!resumeState) {
-        return;
-    }
-
-    if (!quest) {
-        delete resumeQuestIDs[key];
-
-        return;
-    }
-
-    resumeQuestIDs[key] = {
-        timestamp: resumeState.timestamp,
-        questIDs: resumeState.questIDs.filter(id => id !== quest.id),
-    };
 }